# convert-backend/api/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: Convert Gateway API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /api/v1/files/presign:
    post:
      summary: Create a presigned object-storage upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignUploadRequest"
      responses:
        "200":
          description: Upload URL created
  /api/v1/files/{file_id}/complete:
    post:
      summary: Mark an object-storage upload as complete
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File metadata
  /api/v1/files/{file_id}:
    get:
      summary: Get file metadata
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File metadata
        "404":
          description: File not found
  /api/v1/files/{file_id}/download:
    get:
      summary: Create a presigned object-storage download URL
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Download URL created
  /api/v1/jobs:
    post:
      summary: Create an asynchronous conversion job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "202":
          description: Job accepted
  /api/v1/jobs/{job_id}:
    get:
      summary: Get job status
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
components:
  schemas:
    PresignUploadRequest:
      type: object
      required:
        - filename
        - content_type
        - size
      properties:
        user_id:
          type: string
          description: Temporary user identifier until authentication is wired.
        filename:
          type: string
          example: demo.pdf
        content_type:
          type: string
          example: application/pdf
        size:
          type: integer
          format: int64
          example: 1048576
    CreateJobRequest:
      type: object
      required:
        - type
        - input_file_ids
      properties:
        type:
          type: string
          example: image.resize
        input_file_ids:
          type: array
          items:
            type: string
        options:
          type: object
          additionalProperties: true